backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-core
langchain-google-genai
langgraph
google-generativeai
Pillow
aiohttp